% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aha-package.R
\docType{package}
\name{aha-package}
\alias{aha}
\alias{aha-package}
\title{aha: Tools for Statistical Analyis and Teaching}
\description{
The aha package is a set of statistical analysis and teaching tools 
    written by Guy A. Prochilo. Many of these tools have been written to 
    implement a parameter estimation approach to planning research and analyzing
    data, and which have featured in the authors’ empirical work. These tools 
    also serve a teaching purpose, and can be used to clarify introductory 
    statistical concepts. The aha package is named after the interjection 
    'aha!', which refers to the human experience of understanding a previously 
    incomprehensible problem or concept.
}
\author{
\strong{Maintainer}: Guy Prochilo \email{guy.prochilo@gmail.com}

}
\keyword{internal}
