% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.es.R
\name{ci.es}
\alias{ci.es}
\title{Compute Confidence Interval and Margin-of-Error for Cohen's dz,
Cohen's ds, or Pearson's r}
\usage{
ci.es(es, N = NULL, n1 = NULL, n2 = NULL, effect = c("dz", "ds",
  "r"), conf = 0.95, tails = 2, pr = TRUE)
}
\arguments{
\item{es}{Effect size in Cohen's dz, Cohen's ds, or Pearson's r units.}

\item{N}{Total sample size for Cohen's dz or Pearson's r.}

\item{n1}{Sample size for sample 1 (Cohen's ds).}

\item{n2}{Sample size for sample 2 (Cohen's ds).}

\item{effect}{Specify one of "dz", "ds", or "r".}

\item{conf}{Specify the confidence interval width. Defaults to 0.95.}

\item{tails}{Specify whether to perform a 1- or 2-tailed test. Defaults to 2.}

\item{pr}{Specify whether to include printed output. Defaults to TRUE.}
}
\description{
This function computes a confidence interval (CI) and confidence
interval half-width (i.e., margin-of-error; hereby: MoE) for Cohen's dz,
Cohen's ds, or Pearson's r. The CI for Cohen's dz and Cohen's ds is computed
using an exact method based on the noncentral t distribution. The CI for
Pearson's r is computed using Fisher's transformation.

Note: Cohen's dz is computed on paired sample data as the mean difference
divided by the standard deviation of the difference.

Note: Cohen's ds is computed on two sample data as the mean difference
divided by the pooled standard deviation of sample 1 and sample 2.
}
\examples{
# Compute a 2-tailed 95\% CI on Cohen's dz = 0.6 with 20 paired samples
ci.es(es = 0.6, N = 20, effect = "dz", conf = 0.95, tails = 2)

# Compute a 1-tailed 95\% CI on Cohen's dz = 0.6 with 20 paired samples
ci.es(es = 0.6, N = 20, effect = "dz", conf = 0.95, tails = 1)

# Compute a 2-tailed 95\% CI on Cohen's ds = 0.6 with 10 observations in each
# sample
ci.es(es = 0.6, n1 = 10, n2 = 10, effect = "ds", conf = 0.95, tails = 2)

# Compute a 2-tailed 95\% CI on Pearson's r = 0.6 with 20 bivariate samples
ci.es(es = 0.6, N = 20, effect = "r", conf = 0.95, tails = 2)

# Compute a 2-tailed 95\% CI on Pearson's r = 0.6 with 20 bivariate samples,
# but suppress any print statements.
ci.es(es = 0.6, N = 20, effect = "r", conf = 0.95, tails = 2, pr = FALSE)
}
\keyword{Cohen's}
\keyword{Pearson's}
\keyword{coefficient}
\keyword{confidence}
\keyword{correlation}
\keyword{d,}
\keyword{intervals,}
\keyword{r,}
