% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.pearson.r
\name{ci.pearson.r}
\alias{ci.pearson.r}
\title{Confidence interval, margin-of-error, t value, and p value for the
population Pearson's correlation coefficient}
\usage{
ci.pearson.r(r, n, alternative = "two.sided", silent = FALSE)
}
\arguments{
\item{r}{Sample Pearson's correlation coefficient value.}

\item{n}{Total sample size.}

\item{alternative}{Character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{silent}{A logical indicating whether you want the function to return
print output or suppress print output.}
}
\description{
The ci.pearson.r function computes a confidence interval (CI)
and confidence interval half-width (i.e., margin-of-error; hereby: moe) for
the population Pearson's correlation coefficient (r) based on summary
reports of data. If the moe is greater than half the size of r the
function will alert you that this r may have been estimated with low
precision.

For completeness, the function also computes the exact t value and p
value for a given combination of r and sample size (N). This is useful
if the summary report has not included these values.

Note: this function assumes the sample data are outlier free and have a
bivariate normal distribution. Ensure that these assumptions match those of
the reporting publication.
}
\examples{
# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (two-sided test)
ci.pearson.r(r = 0.32, n = 26)

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (hypothesis: the true correlation is greater than 0)
ci.pearson.r(r = 0.49, n = 17, alternative = "greater")

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (hypothesis: the true correlation is less than 0)
ci.pearson.r(r = 0.49, n = 17, alternative = "less")

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (two-sided test) but suppress the print output
ci.pearson.r(r = 0.49, n = 17, silent = TRUE)
}
\keyword{Pearson's}
\keyword{coefficient,}
\keyword{confidence}
\keyword{correlation}
\keyword{intervals}
\keyword{r,}
