% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.r.R
\name{gen.r}
\alias{gen.r}
\title{Generate paired data that gives a specific Pearson correlation}
\usage{
gen.r(r, n, mean.x = 0, mean.y = 0, seed = TRUE, silent = FALSE)
}
\arguments{
\item{r}{Desired Pearson's correlation coefficient value.}

\item{n}{Total sample size.}

\item{mean.x}{Desired mean of x vector}

\item{mean.y}{Desired mean of y vector}

\item{seed}{Set seed = TRUE for reproducible datasets}

\item{silent}{A logical indicating whether you want the function to return
print output or suppress print output.}
}
\description{
The gen.r function generates paired (bivariate) data of a
given sample size with a specific Pearson correlation coefficient. Data are
drawn from a bivariate normal distribution with no outliers.
}
\examples{
# Generate 26 bivariate observations that give a Pearson correlation of 0.32
gendat.r(r = 0.32, n = 26)
}
\keyword{Pearson's}
\keyword{coefficient,}
\keyword{correlation}
\keyword{data}
\keyword{generation}
\keyword{r,}
