% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_ci.R
\name{r_ci}
\alias{r_ci}
\title{Confidence interval, margin-of-error, *t* value, and *p* value for the
population Pearson's correlation coefficient (*r*) based on summary data.}
\usage{
r_ci(r, n, alternative = "two_sided", conf_level = 0.95,
  null_value = 0, silent = FALSE, print_plot = TRUE)
}
\arguments{
\item{r}{Sample Pearson's correlation coefficient value.}

\item{n}{Sample size}

\item{alternative}{Character string specifying the alternative hypothesis,
must be one of "two_sided" (default), "greater" or "less".}

\item{conf_level}{Confidence level of the interval.}

\item{null_value}{The null hypothesis value used to evaluate the sample Pearson's
correlation coefficient (default is 0).}

\item{silent}{A logical indicating whether to suppress message output: FALSE (default) or TRUE.}

\item{print_plot}{A logical indicating whether to print plot of results: FALSE (default) or TRUE.}
}
\description{
The r_ci function computes a confidence interval (CI) and
confidence interval half-width (i.e., margin-of-error; hereby: *MoE*) for
the population Pearson's correlation coefficient (*r*) based on summary
reports of data. If *MoE* is greater than half the size of *r* the function
will alert the user that *r* may have low precision.

For completeness, the function also computes the exact *t* value and *p*
value for a given combination of *r* and sample size (*N*). This is useful
if the summary report has not included these values, or has only included
approximate *p* values (e.g., *p* < .05).

*Important note*: this function assumes the sample data are outlier free and
have a bivariate normal distribution. Ensure that these assumptions match
those of the reporting publication. There will also be a small loss of precision
when using summary data to compute confidence intervals and test statistics.
For example, if you input an *r* value that has been rounded to two decimal
places, the *t* and *p* values may be slightly different from the true values.
These differences, however, will be relatively minor.
}
\examples{
# Compute statistics for a Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (two-sided test)
 r_ci(r = 0.32, n = 26)

# Compute statistics for a Pearson correlation of 0.32 in a sample of 26 bivariate
# observations using the hypothesis that the true correlation is greater than 0:
r_ci(r = 0.32, n = 26, alternative = "greater")

# Compute statistics for a Pearson correlation of 0.32 in a sample of 26 bivariate
# observations using the hypothesis that the true correlation is less than 0:
r_ci(r = 0.32, n = 26, alternative = "less")
}
\keyword{Pearson's}
\keyword{coefficient,}
\keyword{confidence}
\keyword{correlation}
\keyword{intervals}
\keyword{r,}
