% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.pearson.r.R
\name{test.pearson.r}
\alias{test.pearson.r}
\title{Test for Correlation Between Bivariate Samples Using Pearson Correlation
Analysis.}
\usage{
test.pearson.r(x, y, dataset, null = 0, alternative = "two.sided",
  conf.level = 0.95, silent = FALSE)
}
\arguments{
\item{x, y}{numeric vectors of data values. x and y must have the same length.}

\item{dataset}{dataframe that stores x and y vectors}

\item{null}{The null hypothesis value for the Pearson correlation test.
Defaults to 0 if not specified.}

\item{alternative}{Character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{confidence level of the interval.}

\item{silent}{A logical indicating whether you want the function to return
print output or suppress print output.}
}
\description{
The test.pearson.r function performs a Pearson correlation analysis
between bivariate samples. You may specify a specific null hypothesis (defaults
to zero). It returns a test statistic (t for a zero null; z for a nonzero null),
p value, confidence interval, and margin-of-error (confidence interval half-width).
Supports both 1-sided and 2-sided tests, and can report confidence intervals of
different coverage if specified.

Note: this function assumes the sample data are outlier free and have a
bivariate normal distribution.
}
\examples{
test.pearson.r(x = X, y = Y, dataset = df, null = 0.2)
}
\keyword{Pearson's}
\keyword{coefficient,}
\keyword{confidence}
\keyword{correlation}
\keyword{intervals,}
\keyword{nonzero}
\keyword{null}
\keyword{r,}
