% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aha-package.R
\docType{package}
\name{aha-package}
\alias{aha}
\alias{aha-package}
\title{aha: Tools for Statistical Analyis and Teaching}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The aha package is a set of statistical analysis and teaching tools
  that can be used to clarify introductory statistical concepts. The aha package 
  is named after the interjection 'aha!', which refers to the human experience of 
  understanding a previously incomprehensible problem or concept.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gprochilo/aha}
  \item Report bugs at \url{https://github.com/gprochilo/aha/issues}
}

}
\author{
\strong{Maintainer}: Guy Prochilo \email{guy.prochilo@gmail.com}

}
\keyword{internal}
