\name{AHA-package}
\alias{AHA-package}
\alias{AHA}
\docType{package}
\title{
\packageTitle{AHA}
}
\description{
\packageDescription{AHA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AHA}
\packageIndices{AHA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AHA}

Maintainer: \packageMaintainer{AHA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
