% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.pearson.r.R
\name{ci.pearson.r}
\alias{ci.pearson.r}
\title{Confidence interval, margin-of-error, t value, and p value for the
population Pearson's correlation coefficient based on summary data.}
\usage{
ci.pearson.r(r, n, null = 0, alternative = "two.sided",
  conf.level = 0.95, silent = FALSE)
}
\arguments{
\item{r}{Sample Pearson's correlation coefficient value.}

\item{n}{Total sample size.}

\item{null}{The null hypothesis value for the Pearson correlation test.
Defaults to 0 if not specified.}

\item{alternative}{Character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{confidence level of the interval.}

\item{silent}{A logical indicating whether you want the function to return
print output or suppress print output.}
}
\description{
The ci.pearson.r function computes a confidence interval (CI)
and confidence interval half-width (i.e., margin-of-error; hereby: moe) for
the population Pearson's correlation coefficient (r) based on summary
reports of data. If moe is greater than half the size of r the function will
alert you that r may have low precision.

For completeness, the function also computes the exact t value and p
value for a given combination of r and sample size (N). This is useful
if the summary report has not included these values. The user may also
specify a smallest effect size of interest as the null value, if desired,
and compute these statistics based on the alternate hypothesis that the true
correlation is not equal to the null value (this defaults to a null of 0 if
no specified). If a nonzero null is used, a z test is performed and a z
statistic is reported instead of t.

Important note: this function assumes the sample data are outlier free and
have a bivariate normal distribution. Ensure that these assumptions match
those of the reporting publication.

Important note: also be aware that there will be some loss of precision when
using summary data as input. For example, if you input r values that have been
rounded to two decimal places, the computed confidence limit, t, and p values
may be slightly different from the true values. This difference, however, will
be relatively minor.
}
\examples{
# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (two-sided test)
ci.pearson.r(r = 0.32, n = 26)

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (hypothesis: the true correlation is greater than 0)
ci.pearson.r(r = 0.32, n = 26, alternative = "greater")

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (hypothesis: the true correlation is less than 0)
ci.pearson.r(r = 0.32, n = 26, alternative = "less")

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
# observations (two-sided test) but suppress the print output
ci.pearson.r(r = 0.32, n = 26, silent = TRUE)

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
observations testing the null hypothesis that the true correlation is not
equal to 0.5
ci.pearson.r(r = 0.32, n = 26, null = 0.5)

# Examine a reported Pearson correlation of 0.32 in a sample of 26 bivariate
observations and report an 80\% confidence interval
ci.pearson.r(r = 0.32, n = 26, conf.level = 0.8)
}
\keyword{Pearson's}
\keyword{coefficient,}
\keyword{confidence}
\keyword{correlation}
\keyword{intervals,}
\keyword{nonzero}
\keyword{null}
\keyword{r,}
